chore:save

+++ src/pages/database/database-manage/index.tsx
@@ -1,12 +1,13 @@
 import React, { useMemo, useEffect, useState } from 'react';
 import PageContainer from '@/components/page-container';
 import TableSearch from '@/components/table-search';
-import { Button, Space, Form } from 'antd';
+import { Button, Modal, Form, Input, message } from 'antd';
 import useTable from '@/utils/useTable';
 import { createTableColumns } from './schema';
 import CreateDataBase from './create-database';
 import { getSchemaList } from '../service';
 import { useDeleteSchema } from './hook';
+import { ExclamationCircleOutlined } from '@ant-design/icons';
 export interface SchemaProps {
   clusterId: number;
 }
@@ -14,6 +15,7 @@ export interface SchemaProps {
 export default function DEMO(props: SchemaProps) {
   const [form] = Form.useForm();
   const { clusterId } = props;
+  const [ensureForm] = Form.useForm();
   const [mode, setMode] = useState<EditorMode>('HIDE');
   const [delLoading, deleteSchema] = useDeleteSchema();
   useEffect(() => {
@@ -22,13 +24,43 @@ export default function DEMO(props: SchemaProps) {
   const columns = useMemo(() => {
     return createTableColumns({
       onDelete: async (record) => {
-        deleteSchema({ clusterId, id: record?.id }).then(() => {
-          reset();
-        });
+        ensureModal(record);
       },
       delLoading: delLoading,
     }) as any;
   }, []);
+
+  const ensureModal = (record: any) => {
+    Modal.confirm({
+      title: '确定删除该数据库账号吗？',
+      icon: <ExclamationCircleOutlined />,
+      content: (
+        <>
+          <p>
+            您确定要删除此数据库吗？如果是这样，<b>请在此输入数据库名称</b>(
+            <span style={{ color: 'red' }}>{record?.user}</span>)并点击确认删除数据库按钮
+          </p>
+          <Form form={ensureForm}>
+            <Form.Item name="schema">
+              <Input />
+            </Form.Item>
+          </Form>
+          <span>注意：生产环境禁止直接删除数据库！</span>
+        </>
+      ),
+      okText: '确认删除数据库',
+      onOk: async () => {
+        const ensure = record?.name === ensureForm.getFieldsValue()?.schema;
+        if (!ensure) {
+          message.warning('数据库名不一致！');
+        } else {
+          await deleteSchema({ clusterId, id: record?.id }).then(() => {
+            reset();
+          });
+        }
+      },
+    });
+  };
   const {
     tableProps,
     search: { submit, reset },
